ri/backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiokafka==0.13.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.77.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==5.0.1
attrs==25.4.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
black==26.1.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
cssselect2==0.8.0
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.1
elastic-transport==8.17.1
elasticsearch==8.14.0
fastapi==0.128.0
filetype==1.2.0
flake8==7.3.0
fonttools==4.61.1
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.61.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.8
langchain-anthropic==1.3.1
langchain-core==1.2.8
langchain-google-genai==4.2.0
langchain-ollama==1.0.1
langchain-openai==1.1.7
langdetect==1.0.9
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.8
lxml==6.0.2
MarkupSafe==3.0.3
mccabe==0.7.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
ollama==0.6.1
openai==2.16.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
pillow==12.1.0
platformdirs==4.5.1
playwright==1.58.0
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydyf==0.12.1
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
pyphen==0.17.2
PySocks==1.7.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-gnupg==0.5.6
python-jose==3.5.0
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
SQLAlchemy-Utils==0.42.1
starlette==0.50.0
stem==1.8.2
tenacity==9.1.2
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
weasyprint==68.0
webencodings==0.5.1
websockets==15.0.1
wheel==0.46.3
xxhash==3.6.0
yarl==1.22.0
zopfli==0.4.0
zstandard==0.25.0

jinja2
